[refact] repository의 책임 중 file data load 로직 분리
