--- src/main/java/org/prgms/springbootbasic/repository/FileRepository.java
@@ -1,5 +1,0 @@
-package org.prgms.springbootbasic.repository;
-
-public interface FileRepository {
-    void initFile();
-}
