+++ src/main/java/org/prgms/springbootbasic/util/FileManipulator.java
@@ -1,0 +1,5 @@
+package org.prgms.springbootbasic.util;
+
+public interface FileManipulator {
+    void initFile();
+}
